Load/Save map concept
Have a rough working version of saving a map created in the Unity Editor to a file and a way to load it.
Since they both work from start function, it happens simultaneously, so you'll have to disable a few scripts to get it to save/load separately.
Point is that it works.